Fix big-endian test bug: force LE bytes in raw-dtype GGUF test fixture
test_import_raw_dtype_passthrough built its fake GGUF file's F32 payload
with plain ndarray.tobytes(), which serializes in host byte order. GGUF
files are always little-endian on disk, so on a big-endian host (s390x
CI) this produced a non-conformant fixture and the test failed after
round-tripping through the (correctly LE-assuming) loader. Force
little-endian explicitly, matching the struct.pack("<...") convention
already used elsewhere in this file's GGUF writer helpers.

+++ tests/test_import_gguf_weights.py
@@ -235,7 +235,8 @@ def test_import_raw_dtype_passthrough(tmp_path):
     rng = np.random.default_rng(4)
     values = rng.standard_normal(8).astype(np.float32)
     gguf_path = str(tmp_path / "model.gguf")
-    _write_gguf(gguf_path, [("W", GGML_TYPE_F32, [8], values.tobytes())])
+    # GGUF is always little-endian on disk, regardless of host byte order.
+    _write_gguf(gguf_path, [("W", GGML_TYPE_F32, [8], values.astype("<f4").tobytes())])
 
     model = _identity_model("W", [8])
     result, skipped = onnxsim.import_gguf_weights(model, gguf_path)
